requirements.txt
babel==2.17.0
blinker==1.8.2
click==8.1.8
docxcompose==1.4.0
docxtpl==0.16.7
et_xmlfile==2.0.0
Flask==2.3.3
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==2.1.5
numpy==1.24.4
openpyxl==3.1.5
pandas==1.5.3
python-dateutil==2.9.0.post0
python-docx==0.8.11
pytz==2025.1
six==1.17.0
Werkzeug==2.3.7
zipp==3.20.2
